ltp-back/requirements.txt
fastapi[all]==0.95.1
digi-xbee==1.4.1
tomli==2.0.1